wow such euro prices

+++ src/main/java/com/cryptobroz/gnosis_tax_tool/services/CashbackService/dto/DatePrice.java
@@ -1,14 +1,16 @@
 package com.cryptobroz.gnosis_tax_tool.services.CashbackService.dto;
 
+import java.math.BigDecimal;
+
 import com.cryptobroz.gnosis_tax_tool.services.KrakenService.dto.KrakenTickerEntry;
 
 import lombok.Data;
 
 @Data
 public class DatePrice {
-  private Double open;
-  private Double close;
-  private Double avarage;
+  private BigDecimal open;
+  private BigDecimal close;
+  private BigDecimal avarage;
 
   public static DatePrice fromKrakenTickerEntry(final KrakenTickerEntry tickerEntry) {
     DatePrice datePrice = new DatePrice();
